+++ extensions/reclaim-ai/src/utils/events.ts
@@ -1,4 +1,5 @@
 import { ApiResponseEvents } from "../hooks/useEvent.types";
+import { decodeBase32 } from "./base32";
 
 export const eventColors = {
   NONE: "#AAA",
@@ -24,38 +25,6 @@ export const truncateEventSize = (eventTitle: string) => {
   return eventTitle;
 };
 
-const CalendarSyncTitleRegex = /(Personal|Work|Travel) Commitment$/;
-
-/**
- * Group events based on their start-end time combination.
- * @param events List of events to group.
- * @returns Map of grouped events.
- */
-const groupEventsByTime = (events: ApiResponseEvents) => {
-  return events.reduce((acc, event) => {
-    const key = `${event.eventStart}-${event.eventEnd}` as const;
-    if (acc.has(key)) acc.get(key)!.push(event);
-    else acc.set(key, [event]);
-    return acc;
-  }, new Map<string, ApiResponseEvents>());
-};
-
-/**
- * Filter events based on the given criteria.
- * @param events List of events to filter.
- * @returns Filtered list of events.
- */
-const filterRelevantEvents = (events: ApiResponseEvents) => {
-  return events.filter((event) => {
-    const isSyncedAndManaged = event.personalSync && event.reclaimManaged;
-    const matchesTitleCriteria =
-      CalendarSyncTitleRegex.test(event.title) ||
-      (event.title.endsWith("Travel") && events.some((e) => e !== event && e.title.endsWith("Travel")));
-
-    return !(isSyncedAndManaged && matchesTitleCriteria);
-  });
-};
-
 /**
  * Filter out events that are synced, managed by Reclaim and part of multiple calendars
  * @param events
@@ -71,10 +40,21 @@ export function filterMultipleOutDuplicateEvents<Events extends ApiResponseEvent
 ): Events | undefined {
   if (!events) return events;
 
-  // Create a map to store events based on start-end combination
-  const eventTimeMap = groupEventsByTime(events);
+  const ids = new Set(events.map((event) => event.recurringEventId ?? event.eventId));
 
-  return [...eventTimeMap.values()].flatMap((events) =>
-    events.length > 1 ? filterRelevantEvents(events) : events
-  ) as Events;
+  return events.filter((event) => {
+    try {
+      const [type, ...rest] = decodeBase32(event.recurringEventId ?? event.eventId)
+        .replace("\x00", "")
+        .split(":");
+
+      if (type === "reclaim-personal-sync") {
+        const [id] = rest;
+        return id ? !ids.has(id) : true;
+      }
+      return true;
+    } catch (error) {
+      return true;
+    }
+  }) as Events;
 }